Make release script to look in new and old location for openapi.json

+++ release.sh
@@ -67,8 +67,16 @@ MVN_PNC_VERSION=$(mvn -q -Dexec.executable=echo -Dexec.args='${project.version}'
 
 popd && popd
 
-cp pnc/rest-api/pnc-openapi.json .
-check_exit $? "Cannot find or copy pnc-openapi.json, exiting"
+if [ -f pnc/rest-api/pnc-openapi.json ]; then
+    cp pnc/rest-api/pnc-openapi.json .
+fi
+if [ -f pnc/rest-api/target/pnc-openapi.json ]; then
+    cp pnc/rest-api/target/pnc-openapi.json .
+fi
+if ! [ -f ./pnc-openapi.json ]; then
+    echo "Cannot find or copy pnc-openapi.json, exiting"
+    exit 1
+fi
 
 #putting together version for npm package (example: 2.3.0-pnc-2.3.0 or 2.3.0-SNAPSHOT-pnc-aNdosndf)
 NPM_VERSION=(${MVN_PNC_VERSION}-pnc-${PNC_VER})
